Welcome to SAT Collisions
It works now! Just a few minor changes here and there, the major
algorithm was correct. It works on slopes too...just really slow going
up the slope. No jittery effects, no bullets through papers. And there's
a cool glitch where if you jump up to the slope you "hang" onto it, we
could utilize that or fix it.