Made pretty level, added comments, added output for solids.
Level 0 is now a moderately complex level. The idea is that you can
climb to the top and exit slightly faster, with a chance to screw up, or
go through the middle the slower way. Artists: I'm thinking the level
can look like a wide building with rubble where the hole is.
Changing the solidness of a floor will now print a message to the
console that says which state it's in.
Added a comment explaining the usage of solid floors for
HandleEditLevel().