Added mac installer
Inside mac there is a new folder called dmg_creation, it contains
everything used to create the dmg and the output Relo.dmg which can be
used to distribute and install the game on macs. edit background.png to
change the dmg background and create_relo to change other things.
Unfortunately this runs after building so you must wait for it to
create the dmg before you can run it directly from Xamarin.